+++ NEWS.md
@@ -1,3 +1,12 @@
+# actR 0.1.2
+
+* add `gaussianize` options to `detectShiftCore()`, with a default of TRUE, following [@AndreaLemur](https://github.com/AndreaLemur) recognition that that is a requirement for many of the methods in the `changepoint` package. More details on [changepoint workshop here](https://www.youtube.com/watch?v=UfGrLJ7S3sc) 
+
+# actR 0.1.1
+
+* added option to pass plot options to geoChronR plotting to plot.shift()
+* Added better error handling with excursions
+
 # actR 0.1.0
 
 * This marks the first release of the package! Check here for updates